// src/types/user.ts
export interface User {
  id: number;
  first: string;
  last: string;
  email: string;
  phone: string;
  location: string;
  hobby: string;
}

export interface CreateUserData {
  first: string;
  last: string;
  email: string;
  phone: string;
  location: string;
  hobby: string;
}